bb: blinking buttons first imlementation
Configurable functionality (on/off) that connects Led blinking
events with button-backlight led blinking function.
Using LUT_LOOP mode of the PWM functionality of pm8xxx
to repeat the duty cycles till notification is over.
Taking care that the blinking doesnt happen when touchscreen
is on. Also taking into consideration that the blinking
shouldnt interfere with normal working of the key lights.

Author: Illes Pal Zoltan @tbalden